requirements.txt
opencv-python
pillow
pytesseract
pyqt5
tensorflow
tflearn